# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Telemedicina para Todos API",
        "description": "Documentação da API do projeto Telemedicina para Todos. Endpoints internos (sessão) e preparação para API pública de interoperabilidade.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor da aplicação (base atual)"
        },
        {
            "url": "/api/v1",
            "description": "API pública versionada (futuro – interoperabilidade)"
        }
    ],
    "paths": {
        "/api/messages/conversations": {
            "get": {
                "tags": [
                    "Mensagens"
                ],
                "summary": "Listar conversas",
                "description": "Listar conversas do usuário atual",
                "operationId": "94f34a5dbb58ee0bcc1a7908ed356c2a",
                "responses": {
                    "200": {
                        "description": "Lista de conversas"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
        "/api/messages/{userId}": {
            "get": {
                "tags": [
                    "Mensagens"
                ],
                "summary": "Mensagens com um usuário",
                "description": "Buscar mensagens entre o usuário atual e outro usuário",
                "operationId": "9cb6e378b77d6b632dc1b89c5d24dc50",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID do outro usuário",
                        "required": true
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "before",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de mensagens"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
        "/api/messages": {
            "post": {
                "tags": [
                    "Mensagens"
                ],
                "summary": "Enviar mensagem",
                "description": "Enviar uma mensagem",
                "operationId": "c85a9171473a13f151a8672235e01559",
                "responses": {
                    "201": {
                        "description": "Mensagem enviada"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/api/messages/{userId}/read": {
            "post": {
                "tags": [
                    "Mensagens"
                ],
                "summary": "Marcar mensagens como lidas",
                "description": "Marcar mensagens como lidas",
                "operationId": "83ffd2c44909298da45e30c019f9fa32",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID do usuário remetente",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensagens marcadas como lidas"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
...
        "/api/messages/unread/count": {
            "get": {
                "tags": [
                    "Mensagens"
                ],
                "summary": "Contagem de mensagens não lidas",
                "description": "Contar mensagens não lidas",
                "operationId": "beb150d58e2cec57073098e7fa85e3d2",
                "responses": {
                    "200": {
                        "description": "Contagem"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
        "/api/appointments/availability": {
            "get": {
                "tags": [
                    "Agendamentos"
                ],
                "summary": "Disponibilidade para agendamento",
                "description": "Retorna horários disponíveis do médico na data. Requer autenticação (sessão).",
                "operationId": "8429ea764d98d0a9b098d603a8586bac",
                "parameters": [
                    {
                        "name": "doctor_id",
                        "in": "query",
                        "description": "ID do médico",
                        "required": true
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Data (Y-m-d)",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Slots disponíveis e agenda do dia"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Médico inativo ou dados inválidos"
                    }
                }
            }
        },
...
    },
...
}